fix: some final fixes
